+++ effect_visit_schedule/visit_schedules/crfs.py
@@ -41,7 +41,7 @@ crfs_d01 = FormsCollection(
     Crf(show_order=60, model="effect_subject.diagnoses"),
     Crf(show_order=70, model="effect_subject.chestxray"),
     Crf(show_order=80, model="effect_subject.arvtreatment"),
-    Crf(show_order=90, model="effect_subject.studytreatment"),
+    Crf(show_order=90, model="effect_subject.patienttreatment"),
     Crf(show_order=100, model="effect_subject.bloodresultsfbc"),
     Crf(show_order=140, model="effect_subject.lpcsf"),
     Crf(show_order=160, model="effect_subject.microbiology"),